requirements.txt
django-tenant-schemas